move rpcclient and types package to common
