+++ cmd/client/client/discover.go
@@ -4,8 +4,8 @@ import (
 	"encoding/json"
 
 	"github.com/noot/atomic-swap/common"
+	"github.com/noot/atomic-swap/common/rpcclient"
 	"github.com/noot/atomic-swap/rpc"
-	"github.com/noot/atomic-swap/rpcclient"
 )
 
 // Discover calls net_discover.
