+++ common/types/offer.go
