+++ monero/client.go
@@ -2,8 +2,8 @@ package monero
 
 import (
 	"github.com/noot/atomic-swap/common"
+	"github.com/noot/atomic-swap/common/rpcclient"
 	mcrypto "github.com/noot/atomic-swap/crypto/monero"
-	"github.com/noot/atomic-swap/rpcclient"
 )
 
 // Client represents a monero-wallet-rpc client.
