+++ rpc/server.go
@@ -5,8 +5,8 @@ import (
 	"net/http"
 
 	"github.com/noot/atomic-swap/common"
+	"github.com/noot/atomic-swap/common/types"
 	"github.com/noot/atomic-swap/net"
-	"github.com/noot/atomic-swap/types"
 
 	"github.com/gorilla/mux"
 	"github.com/gorilla/rpc/v2"
